+++ packages/expo-widgets/src/index.ts
@@ -3,6 +3,7 @@ export type {
   ExpoWidgetsEvents,
   LevelOfDetail,
   LiveActivityComponent,
+  LiveActivityContentState,
   LiveActivityDismissalPolicy,
   LiveActivityEnvironment,
   LiveActivityEvents,